# Exponentially faster tree depth