@pbiecek 2024-03-13
binary search wins!
thanks